+++ tests/snowpark/test_snowpark_shared.py
@@ -4,8 +4,7 @@ from unittest import mock
 from zipfile import ZipFile
 
 import snowflake.cli.plugins.snowpark.snowpark_shared as shared
-from requirements.requirement import Requirement
-from snowflake.cli.plugins.snowpark.models import SplitRequirements
+from snowflake.cli.plugins.snowpark.models import Requirement, SplitRequirements
 
 
 @mock.patch("snowflake.cli.plugins.snowpark.package_utils.parse_anaconda_packages")